Incorporated the use HeaderClickListener in my activities. Updated use of ProductPage with User and Product as extras. Getting rid of the infamous ClothingItem.java for a Product, with all changes depending on that. Ordering GiveItemActivity code so it flows better when reading.

+++ app/src/main/java/com/benefit/utilities/staticClasses/Converters.java
@@ -34,7 +34,7 @@ public class Converters {
         return false;
     }
 
-    public static int getMetaCategoryFromCategoryId(int categoryId) {
+    public static int getMetaCategoryFromCategoryId(long categoryId) {
         if (categoryId == 2 || categoryId == 3) {
             return 1;
         } else if (categoryId == 5 || categoryId == 6){
